requirements.txt
absl-py~=0.13.0
arabic-reshaper~=2.1.1
argon2-cffi~=21.1.0
astunparse~=1.6.3
attrs~=21.2.0
backcall~=0.2.0
bleach~=4.1.0
cachetools~=4.2.2
certifi~=2021.5.30
cffi~=1.14.6
charset-normalizer~=2.0.4
cloudpickle~=2.0.0
cycler~=0.10.0
Cython~=0.29.22
dask~=2021.9.0
debugpy~=1.4.3
decorator~=5.1.0
defusedxml~=0.7.1
entrypoints~=0.3
flatbuffers~=1.12
fsspec~=2021.8.1
func-timeout~=4.3.5
future~=0.18.2
gast~=0.4.0
geneticalgorithm~=1.0.2
google-auth~=1.34.0
google-auth-oauthlib~=0.4.5
google-pasta~=0.2.0
grpcio~=1.34.1
h5py~=3.1.0
idna~=3.2
ipykernel~=6.4.1
ipython~=7.27.0
ipython-genutils~=0.2.0
ipywidgets~=7.6.4
jdatetime~=3.6.2
jedi~=0.18.0
Jinja2~=3.0.1
joblib~=1.0.1
jsonschema~=3.2.0
jupyter-client~=7.0.2
jupyter-core~=4.7.1
jupyterlab-pygments~=0.1.2
jupyterlab-widgets~=1.0.1
k-means-constrained~=0.6.0
keras-nightly~=2.5.0.dev2021032900
Keras-Preprocessing~=1.1.2
kiwisolver~=1.3.1
llvmlite~=0.36.0
locket~=0.2.1
Markdown~=3.3.4
MarkupSafe~=2.0.1
matplotlib~=3.3.4
matplotlib-inline~=0.1.3
mistune~=0.8.4
nbclient~=0.5.4
nbconvert~=6.1.0
nbformat~=5.1.3
nest-asyncio~=1.5.1
notebook~=6.4.3
numba~=0.53.1
oauthlib~=3.1.1
opt-einsum~=3.3.0
ortools~=9.1.9490
packaging~=21.0
pandas~=1.1.5
pandocfilters~=1.4.3
parso~=0.8.2
partd~=1.2.0
persiantools~=2.1.1
pexpect~=4.8.0
pickleshare~=0.7.5
Pillow~=8.1.0
prometheus-client~=0.11.0
prompt-toolkit~=3.0.20
protobuf~=3.18.0
psutil~=5.8.0
ptyprocess~=0.7.0
py4j~=0.10.7
pyasn1~=0.4.8
pyasn1-modules~=0.2.8
pycparser~=2.20
Pygments~=2.10.0
pypandoc~=1.5
pyparsing~=2.4.7
PyQt5~=5.15.4
PyQt5-Qt5~=5.15.2
PyQt5-sip~=12.8.1
PyQt5-stubs~=5.15.2.0
pyrsistent~=0.18.0
python-bidi~=0.4.2
python-dateutil~=2.8.1
pytz~=2021.1
PyWavelets~=1.1.1
PyYAML~=5.4.1
pyzmq~=22.2.1
requests~=2.26.0
requests-oauthlib~=1.3.0
rsa~=4.7.2
scikit-learn~=0.24.1
scipy~=1.7.1
seaborn~=0.11.1
Send2Trash~=1.8.0
six~=1.15.0
stumpy~=1.8.0
swifter~=1.0.9
tensorboard~=2.6.0
tensorboard-data-server~=0.6.1
tensorboard-plugin-wit~=1.8.0
tensorflow~=2.5.1
tensorflow-estimator~=2.5.0
termcolor~=1.1.0
terminado~=0.12.1
testpath~=0.5.0
threadpoolctl~=2.1.0
toolz~=0.11.1
tornado~=6.1
tqdm~=4.62.2
traitlets~=5.1.0
tslearn~=0.5.0.5
typing-extensions~=3.7.4.3
urllib3~=1.26.6
wcwidth~=0.2.5
webencodings~=0.5.1
Werkzeug~=2.0.1
widgetsnbextension~=3.5.1
wrapt~=1.12.1
xlrd~=1.1.0
XlsxWriter~=1.3.7
numpy~=1.21.2